+++ backend/src/main/java/github/nettopro/rpgenzo/entidade/acao/AcaoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AcaoRepository extends JpaRepository<Acao, Long> {
+public interface AcaoRepository extends JpaRepository<Acao, Integer> {
 
     Optional<Acao> findByNomeIgnoreCase(String nome);
 }
